Refactor: 컴포넌트 파일 재분리
